+++ FB6.cpp
@@ -18,7 +18,7 @@ FB6::FB6() : kappa(1), beta(1), gamma(1)
 /*!
  *  Constructor
  */
-FB6::FB6(long double kappa, long double beta, long double gamma) : 
+FB6::FB6(double kappa, double beta, double gamma) : 
         kappa(kappa), beta(beta), gamma(gamma)
 {
   mu = ZAXIS;
@@ -30,7 +30,7 @@ FB6::FB6(long double kappa, long double beta, long double gamma) :
  *  Constructor
  */
 FB6::FB6(Vector &mu, Vector &major_axis, Vector &minor_axis,
-         long double kappa, long double beta, long double gamma) :
+         double kappa, double beta, double gamma) :
          mu(mu), major_axis(major_axis), minor_axis(minor_axis), 
          kappa(kappa), beta(beta), gamma(gamma)
 {}
@@ -67,19 +67,19 @@ std::vector<Vector> FB6::generate(int sample_size)
 std::vector<Vector> FB6::generateCanonical(int sample_size)
 {
   if (beta != 0) {  // general case
-    long double psi1 = gamma - beta;
-    long double psi2 = gamma + beta;
+    double psi1 = gamma - beta;
+    double psi2 = gamma + beta;
     FB4 g1(kappa,psi1);
     FB4 g2(kappa,psi2);
-    long double c1 = g1.computeNormalizationConstant();
-    long double c2 = g2.computeNormalizationConstant();
-    long double num,denom;
+    double c1 = g1.computeNormalizationConstant();
+    double c2 = g2.computeNormalizationConstant();
+    double num,denom;
     Vector u(sample_size,0);
     // step 1: FB6_0
     // step 0
     denom = c1 + exp(-2*beta) * c2;
-    long double p2 = c1 / denom;
-    long double s1,s2,eta,tmp;
+    double p2 = c1 / denom;
+    double s1,s2,eta,tmp;
     Vector u1;
     for (int i=0; i<sample_size; i++) {
       // step 1
@@ -105,7 +105,7 @@ std::vector<Vector> FB6::generateCanonical(int sample_size)
     } // for loop ends ...
     // step 2: FB6_0
     Vector phi(sample_size,0);
-    long double p,psi,angle;
+    double p,psi,angle;
     int delta;
     Vector vmf2dmean(2,0); vmf2dmean[0] = 1; 
     std::vector<Vector> random_sample;
@@ -168,7 +168,7 @@ std::vector<Vector> FB6::generate_cartesian_coordinates(Vector &u, Vector &phi)
 {
   std::vector<Vector> coordinates(u.size());
   Vector x(3,0);
-  long double tmp;
+  double tmp;
   for (int i=0; i<u.size(); i++) {
     tmp = sqrt(1-u[i]*u[i]);
     x[0] = tmp * cos(phi[i]);
