+++ Normal.cpp
@@ -11,10 +11,10 @@ Normal::Normal() : mu(0), sigma(1)
 /*!
  *  \brief constructor function which sets the value of mean and 
  *  standard deviation of the distribution
- *  \param mu a long double
- *  \param sigma a long double
+ *  \param mu a double
+ *  \param sigma a double
  */
-Normal::Normal(long double mu, long double sigma) : mu(mu), sigma(sigma)
+Normal::Normal(double mu, double sigma) : mu(mu), sigma(sigma)
 {}
 
 /*!
@@ -34,7 +34,7 @@ Normal Normal::operator=(const Normal &source)
  *  \brief This function returns the mean of the distribution
  *  \return the mean of the distribution
  */
-const long double Normal::mean(void)
+const double Normal::mean(void)
 {
 	return mu;
 }
@@ -43,31 +43,31 @@ const long double Normal::mean(void)
  *  \brief This function returns the standard deviation of the distribution
  *  \return the standard deviation of the distribution
  */
-const long double Normal::standardDeviation(void)
+const double Normal::standardDeviation(void)
 {
 	return sigma;
 }
 
 /*!
  *  \brief This function computes the value of the distribution at a given x
- *  \param x a long double
+ *  \param x a double
  *  \return density of the function given x
  */
-long double Normal::density(long double x)
+double Normal::density(double x)
 {
-  long double tmp = (x-mu)/(long double)sigma;
-  long double exponent = -0.5 * tmp * tmp;
-	long double val = exp(exponent) / ((sqrt(2*PI)) * sigma);
+  double tmp = (x-mu)/(double)sigma;
+  double exponent = -0.5 * tmp * tmp;
+	double val = exp(exponent) / ((sqrt(2*PI)) * sigma);
   return val;
 }
 
 /*!
  *  cdf = \int_{-\infty}^x f(x) dx
  */
-long double Normal::cumulativeDensity(long double x)
+double Normal::cumulativeDensity(double x)
 {
-  long double z = (mu - x) / (sqrt(2) * sigma);
-  long double error = erfc(z);
+  double z = (mu - x) / (sqrt(2) * sigma);
+  double error = erfc(z);
   return error/2;
 }
 
@@ -76,9 +76,9 @@ long double Normal::cumulativeDensity(long double x)
  *  \param sample a reference to a Vector
  *  \return the negative log likelihood (base e)
  */
-long double Normal::negativeLogLikelihood(long double x)
+double Normal::negativeLogLikelihood(double x)
 {
-  long double pdf = density(x);
+  double pdf = density(x);
   return -log(pdf);
 }
 
@@ -87,18 +87,18 @@ long double Normal::negativeLogLikelihood(long double x)
  *  \param sample a reference to a Vector
  *  \return the negative log likelihood (base e)
  */
-long double Normal::negativeLogLikelihood(Vector &sample)
+double Normal::negativeLogLikelihood(Vector &sample)
 {
-  long double value = 0;
-  long double tmp = 0.5 * log(2*PI) + log(sigma);
+  double value = 0;
+  double tmp = 0.5 * log(2*PI) + log(sigma);
   value += (sample.size() * tmp);
 
-  long double num = 0;
+  double num = 0;
   for (int i=0; i<sample.size(); i++) {
     num += (sample[i]-mu) * (sample[i]-mu);
   }
-  long double denom = 2 * sigma * sigma;
-  value += num / (long double)denom;
+  double denom = 2 * sigma * sigma;
+  value += num / (double)denom;
   return value;
 }
 
@@ -119,7 +119,7 @@ void Normal::printParameters(ostream &os)
 Vector Normal::generate(int sample_size)
 {
   Vector sample(sample_size,0);
-  long double u,v,r1,r2,sqroot,arg;
+  double u,v,r1,r2,sqroot,arg;
 
   for (int i=0; i<sample_size; i+=2) {
     repeat:
